Move the code created by LG-6797
Move code to a different location in our codebase so that it has
a more appropriate namespace.

[skip changelog]
