+++ spec/support/shared_contexts/inherited_proofing/va_api_context.rb
@@ -9,7 +9,7 @@ RSpec.shared_context 'va_api_context' do
   let(:payload) { { inherited_proofing_auth: auth_code, exp: 1.day.from_now.to_i } }
   let(:jwt_token) { JWT.encode(payload, private_key, 'RS256') }
   let(:request_uri) {
-    "#{InheritedProofing::Va::Service::BASE_URI}/inherited_proofing/user_attributes"
+    "#{Idv::InheritedProofing::Va::Service::BASE_URI}/inherited_proofing/user_attributes"
   }
   let(:request_headers) { { Authorization: "Bearer #{jwt_token}" } }
 end
